Let’s not have randomstrseed.
Because of Lua’s built-in coercion, we can have math.randomseed simply
take a string argument.  If one gives randomseed() a number, Lua is
smart enough to perform coercion: It makes the number a string so
that the API both is backwards compatible, and allows us to have
string random seeds.