mmLunacyDNS: We now use programName.lua as config file
So, if the binary is called mmLunacyDNS.exe or mmLunacyDNS, we
use mmLunacyDNS.lua as the config file.  If we rename it foo.exe (or foo,
or foo.bar), we use foo.lua as the config file.  The rule is this:
Whatever name we give the executable is the name we give the .lua
config file.