mmLunacyDNS: We now have a well defined, if simple, API
mmLunacyDNS reads the Lua configuration file.  From that file,
it gets the string bindIp, which is the IP mmLunacyDNS binds to.

Once it binds to the IP, every time mmLunacyDNS gets a query, it
runs the lua function processQuery, which takes as its input a
table with two members:

--

mmQuery: This is the DNS name requested, in the form of a string
like "caulixtla.com." or "samiam.org.".  If the string has anything
besides an ASCII letter, an ASCII number, the '-' character (dash), or
the '_' character (underline), the character will be a two-digit
hexadecimal number in brackets.  If we get the raw UTF-8 query
"ñ.samiam.org" (where the first character is a n with a tilde),
mmQuery will look like "{c3}{b1}.samiam.org.".

mmQtype: The is the numeric DNS query type requested.  This is a number
between 0 and 65535, and corresponds to the DNS query type made.  A
list of DNS query type numbers is available at
https://www.iana.org/assignments/dns-parameters/dns-parameters.xhtml
1 is "A", i.e. a request for an IPv4 IP address.

--

The processQuery function returns as its output a table with two
parameters:

mm1Type: This right now has to be the string "A", or the DNS query is
ignored

mm1Data: This is an IPv4 IP in dotted decimal format, e.g. "10.1.2.3"

--

Simple example:

-- We bind to the IP 127.0.0.1 and always return "10.1.2.3" for every
-- DNS query we get
bindIp = "127.0.0.1"
function processQuery(mmAll)
  return {mm1Type = "A", mm1Data="10.1.2.3"}
end

--

More complex example, where .com gets 10.1.1.1 and anything else
gets 10.1.2.3, non-A queries are ignored, and we log A queries:

bindIp = "127.0.0.1"
function processQuery(mmAll)
  if mmAll.mmQtype ~= 1 then
    return {mm1Type = "ignoreMe"}
  end
  mmDNS.log("Got IPv4 query for " .. mmAll.mmQuery)
  if string.match(mmAll.mmQuery,'.com.$') then
    return {mm1Type = "A", mm1Data = "10.1.1.1"}
  end
  return {mm1Type = "A", mm1Data = "10.1.2.3"}
end