Lunacy 2021-03-21 Better desktop calculator
When in interactive mode, if the first character is a number, we will
return the results of the arithmetic expression.  This way, if one types
'1 + 1' (w/o quotes) while in terminal mode, Lunacy will print “1” on the
terminal.

For things like math.sin(3.141592653 / 4), '=' is still needed, e.g.
'= math.sin(3.141592653 / 4)', but [0-9] covers the lion’s share of
cases when I need a simple “desktop calculator”.