COPYING.md (LUAstuff): Change "license", such as it is, to unlicense
The reason to use unlicense instead of CC0 is because CC0 implies
the code might be patented.  Unlicense implies no such thing.  For
the record, anything I dedicate to the public domain does not have,
to the extent of my knowledge, the use of any patents in the
algorithms or procedures utilized by the code.