Just have lunacy.today() fail in systems with 32-bit time_t
It’s too close to the Y2038 point to allow code which fails on
that day to be around, so just have lunacy.today() fail if time_t
is 32 bits in size, with notes to use https://github.com/evalEmpire/y2038
for people who must have this work after Y2038 and can not make
time_t bigger.  The Linux kernel already allows one to get a
64-bit timestamp on a 32-bit system, and I believe there are libc
implementations which do so too.